修改购买会员(新,续费)

+++ database/migrations/2017_01_08_212921_create_orders_table.php
@@ -14,7 +14,7 @@ class CreateOrdersTable extends Migration
 	public function up()
 	{
 		Schema::create('orders', function(Blueprint $table) {
-            $table->integer('id')->unsigned()->comment('订单id');
+            $table->bigInteger('id')->unsigned()->comment('订单id');
             $table->decimal('order_amount')->unsigned()->default(0.00)->comment('订单总金额');
             $table->decimal('pay_amount')->unsigned()->default(0.00)->comment('应付总金额');
             $table->enum('pay_method', ['alipay','wechat'])->comment('支付方式');
